+++ RELEASE_NOTES.md
@@ -19,6 +19,7 @@ _16 June 2016_
 * Fixed missing "N/A" in Cert ID PDF
 * Fixed Cert ID Lookup results header always showing
 * Fixed widget product list buttons size 
+* Fixed Cert ID format checking for Cert ID Lookup input field
 
 ---